api/requirements.txt
uvicorn==0.32.1
fastapi==0.115.5
requests==2.31.0
pymongo==4.10.1
python-dotenv==1.0.1